pyproject.toml
[project]
name = "smartstitch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.3",
    "pillow>=12.1.1",
    "tqdm>=4.67.3",
]
